LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/radicals/glyph-level-mirroring-expected.txt


FAIL LTR Direction assert_approx_equals: Base glyph expected 80 +/- 1 but got 24.78125
FAIL RTL Direction assert_approx_equals: Base glyph expected 32 +/- 1 but got 24.78125
